Fixed linux compile
git-svn-id: http://svn.kulitorum.com/RepSnapper/Src@146 686d4eb6-aa70-4f30-b8fa-ee57029aed71

+++ Flu_Tree_Browser.cpp
@@ -10,9 +10,9 @@
  * Attn: Jason Bryan Re: FLU 1224 Kinnear Rd, Columbus, Ohio 43212
  * 
  ***************************************************************/
-
-
-
+/* davekw7x: Changed Lines 2984, 2986, 2989, 2999, 3004, 3031, 3036
+ * from unsigned int to unsigned long to make it compatable with g++ 
+ * version 4 64-bit compiler */
 #include <FL/Fl.H>
 #include <FL/fl_draw.H>
 #include <FL/math.h>
@@ -2981,12 +2981,12 @@ Flu_Tree_Browser::Node* Flu_Tree_Browser :: add_leaf( const char* path, const ch
   return add( path, p.c_str(), w, showLabel );
 }
 
-unsigned int Flu_Tree_Browser :: remove( const char *fullpath )
+unsigned long Flu_Tree_Browser :: remove( const char *fullpath )
 {
-  return( (unsigned int)root.modify( fullpath, Node::REMOVE, rdata ) );
+  return( (unsigned long )root.modify( fullpath, Node::REMOVE, rdata ) );
 }
 
-unsigned int Flu_Tree_Browser :: remove( const char *path, const char *text )
+unsigned long Flu_Tree_Browser :: remove( const char *path, const char *text )
 {
   // if the path does not end in '/', add it
   FluSimpleString s = path;
@@ -2996,12 +2996,12 @@ unsigned int Flu_Tree_Browser :: remove( const char *path, const char *text )
   return remove( s.c_str() );
 }
 
-unsigned int Flu_Tree_Browser :: remove( unsigned int id )
+unsigned long Flu_Tree_Browser :: remove( unsigned int id )
 {
   return root.remove( id );
 }
 
-unsigned int Flu_Tree_Browser :: Node :: remove( unsigned int id )
+unsigned long Flu_Tree_Browser :: Node :: remove( unsigned int id )
 {
   if( id == 0 )
     return 0;
@@ -3028,12 +3028,12 @@ unsigned int Flu_Tree_Browser :: Node :: remove( unsigned int id )
   return 0;
 }
 
-unsigned int Flu_Tree_Browser :: remove( Fl_Widget *w )
+unsigned long Flu_Tree_Browser :: remove( Fl_Widget *w )
 {
   return root.remove( w );
 }
 
-unsigned int Flu_Tree_Browser :: Node :: remove( Fl_Widget *w )
+unsigned long Flu_Tree_Browser :: Node :: remove( Fl_Widget *w )
 {
   if( !w )
     return 0;
